64-bit compatibility
We can (and will) now build for 64-bit systems! \o/

Some interesting things were needed for this to happen:

- There was a "typedef unsigned long ADDR" for a type which
  is used to cast pointers from; fine on LP64 systems like
  Linux, but catastrophic in Win32's LLP64 environment.
  This has been replaced by uintptr_t.

- For reasons I don't fully understand, the LZW decoder
  overreads a few bytes in the very last file. (This happens
  consistently on Windows and Linux.) This puts the .dat file
  reader (which expects clients to read their files *exactly*
  until the end, no more, no less) waaay off its tracks.
  A few safeguards have been added to cause a "rewind" of the
  .dat file if the filename/filesize headers stop making sense.

- A weird crash on Win32/x86_64 happened because I forgot to
  add those stupid '#pragma pack' directives to a structure
  which inherits from another one that *is* packed. Fun
  ensued when casting the one to the other ...

+++ include/main.h
@@ -6,6 +6,8 @@
 #ifndef _MAIN_H_
 #define _MAIN_H_
 
+#include <stdint.h>
+
 #define EXTERN
 
 #ifndef __WATCOMC__
@@ -35,7 +37,7 @@ typedef float              FLT;
 
 typedef unsigned int       UINT;
 typedef int                INT;
-typedef unsigned long int  ADDR;
+typedef uintptr_t          ADDR;
 
    // for the gfx lib...
 typedef PIXEL COLOR_ENTRY[4];
